each concrete aggregator has it's own file now
